+++ packages/mosaic/__tests__/ChartSettings.test.tsx
@@ -49,7 +49,7 @@ describe('MosaicChartSettings Compound Components', () => {
         'line-chart',
         'heatmap',
         'box-plot',
-        'scatter-plot-chart',
+        'scatter-plot',
       ];
 
       chartTypes.forEach((chartType) => {
