+++ packages/mosaic/__tests__/dashboard-ai.test.ts
@@ -341,7 +341,7 @@ describe('dashboard AI tools', () => {
               type: 'vgplot',
               title: 'Magnitude vs depth',
               config: {
-                chartType: 'scatter-plot-chart',
+                chartType: 'scatter-plot',
                 settings: {x: 'magnitude', y: 'depth'},
               },
             },
@@ -352,7 +352,7 @@ describe('dashboard AI tools', () => {
         [`${dashboardId}:${panelId}`]: {
           kind: 'too-much-data',
           panelId,
-          chartType: 'scatter-plot-chart',
+          chartType: 'scatter-plot',
           message: 'Use a heatmap instead.',
           recoverable: true,
           rowCount: 20,
@@ -370,7 +370,7 @@ describe('dashboard AI tools', () => {
     expect(result.llmResult.success).toBe(true);
     expect(result.llmResult.data.panels[0].issue).toMatchObject({
       kind: 'too-much-data',
-      chartType: 'scatter-plot-chart',
+      chartType: 'scatter-plot',
       rowCount: 20,
       limit: 10,
     });
