+++ packages/mosaic/src/charts/chart-types/index.ts
@@ -17,7 +17,7 @@ export * from './line-chart/schema';
 export * from './count-plot/schema';
 export * from './heatmap/schema';
 export * from './box-plot/schema';
-export * from './scatter-plot-chart/schema';
+export * from './scatter-plot/schema';
 export * from './custom-spec/schema';
 
 // Re-export Settings components
@@ -26,7 +26,7 @@ export {LineChartSettingsComponent} from './line-chart/LineChartSettings';
 export {CountPlotSettingsComponent} from './count-plot/CountPlotSettings';
 export {HeatmapSettingsComponent} from './heatmap/HeatmapSettings';
 export {BoxPlotSettingsComponent} from './box-plot/BoxPlotSettings';
-export {ScatterPlotSettingsComponent} from './scatter-plot-chart/ScatterPlotSettings';
+export {ScatterPlotSettingsComponent} from './scatter-plot/ScatterPlotSettings';
 export {CustomSpecSettingsComponent} from './custom-spec/CustomSpecSettings';
 
 // Re-export definitions
@@ -35,7 +35,7 @@ export * from './line-chart/definition';
 export * from './count-plot/definition';
 export * from './heatmap/definition';
 export * from './box-plot/definition';
-export * from './scatter-plot-chart/definition';
+export * from './scatter-plot/definition';
 export * from './custom-spec/definition';
 
 // Re-export tool types, schemas, validation, helpers, and AI tool creators
@@ -47,7 +47,7 @@ export * from './histogram/tool';
 export * from './line-chart/tool';
 export * from './count-plot/tool';
 export * from './heatmap/tool';
-export * from './scatter-plot-chart/tool';
+export * from './scatter-plot/tool';
 export * from './box-plot/tool';
 export * from '../../ai/data-table-explorer-tool';
 export * from '../../ai/list-panels-tool';
