+++ packages/mosaic/src/charts/chart-types/scatter-plot/spec.ts
